source3/passdb: Follow up to the fix for CID 1508970
Fixing another occurrence for the same CID in the same DEBUG statement.
See the previous commit c25a5a26a1cba698420fb64e23bee3b52540b21b.


Autobuild-User(master): Anoop C S <anoopcs@samba.org>
Autobuild-Date(master): Wed Sep 10 15:23:08 UTC 2025 on atb-devel-224

+++ source3/passdb/login_cache.c
@@ -116,8 +116,8 @@ bool login_cache_read(struct samu *sampass, struct login_cache *entry)
 
 	SAFE_FREE(databuf.dptr);
 
-	DEBUG(5, ("Found login cache entry: timestamp %12u, flags 0x%x, count %d, time %12jd\n",
-		  (unsigned int)entry->entry_timestamp, entry->acct_ctrl, 
+	DEBUG(5, ("Found login cache entry: timestamp %12jd, flags 0x%x, count %d, time %12jd\n",
+		  (intmax_t)entry->entry_timestamp, entry->acct_ctrl,
 		  entry->bad_password_count, (intmax_t)entry->bad_password_time));
 	return true;
 }
